added field to translations data frame
