+++ isovar/translation.py
@@ -586,8 +586,20 @@ def translations_generator_to_dataframe(translations_generator):
     return dataframe_from_generator(
         element_class=Translation,
         variant_and_elements_generator=translations_generator,
-        # exlude fields which are structured objects
-        exclude=[
-            "variant_sequence",
-            "reference_context",
-            "variant_sequence_in_reading_frame"])
+        exclude=["reference_context"],
+        converters={
+            "untrimmed_variant_sequence": lambda vs: vs.sequence,
+            "variant_sequence_in_reading_frame": lambda vs: vs.sequence,
+            "reference_context": lambda rc: rc,
+        },
+        extra_column_fns={
+            "untrimmed_variant_sequence_read_count": (
+                lambda t: len(t.untrimmed_variant_sequence.reads)),
+            "variant_sequence_in_reading_frame_read_count": (
+                lambda t: len(t.variant_sequence_in_reading_frame.reads)),
+            "reference_context_transcripts": (
+                lambda t: ";".join([
+                    transcript.name for
+                    transcript in
+                    t.reference_context.transcripts]))
+        })
